Product CRUD

+++ database/migrations/2024_12_07_073005_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->bigInteger('price');
             $table->longText('description');
-            $table->integer('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->integer('stock');
             $table->string('image');
             $table->timestamps();
